+++ braintree/braintree_gateway.py
@@ -17,6 +17,7 @@ from braintree.oauth_gateway import OAuthGateway
 from braintree.payment_method_gateway import PaymentMethodGateway
 from braintree.payment_method_nonce_gateway import PaymentMethodNonceGateway
 from braintree.paypal_account_gateway import PayPalAccountGateway
+from braintree.paypal_payment_resource_gateway import PayPalPaymentResourceGateway
 from braintree.sepa_direct_debit_account_gateway import SepaDirectDebitAccountGateway
 from braintree.plan_gateway import PlanGateway
 from braintree.settlement_batch_summary_gateway import SettlementBatchSummaryGateway
@@ -60,6 +61,7 @@ class BraintreeGateway(object):
         self.payment_method = PaymentMethodGateway(self)
         self.payment_method_nonce = PaymentMethodNonceGateway(self)
         self.paypal_account = PayPalAccountGateway(self)
+        self.paypal_payment_resource = PayPalPaymentResourceGateway(self)
         self.plan = PlanGateway(self)
         self.sepa_direct_debit_account = SepaDirectDebitAccountGateway(self)
         self.settlement_batch_summary = SettlementBatchSummaryGateway(self)