+++ braintree/test/credit_card_numbers.py
@@ -6,6 +6,7 @@ class CreditCardNumbers(object):
         Healthcare = "4111111510101010"
         Payroll  = "4111111114101010"
         Prepaid = "4111111111111210"
+        PrepaidReloadable = "4229989900000002"
         IssuingBank = "4111111141010101"
         CountryOfIssuance = "4111111111121102"